[#675] Refactor: endpoint generate-session implementado con arquitectura por capas
