+++ AuthMicroservice Backend/src/main/java/upb/edu/AuthMicroservice/controllers/SessionController.java
@@ -1,37 +1,30 @@
 package upb.edu.AuthMicroservice.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import upb.edu.AuthMicroservice.models.Session;
-import upb.edu.AuthMicroservice.models.GenerateSessionRequest;
-import upb.edu.AuthMicroservice.repositories.SessionRepository;
+import upb.edu.AuthMicroservice.dtos.GenerateSessionRequest;
+import upb.edu.AuthMicroservice.services.SessionService;
 
-import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/")
 public class SessionController {
 
     @Autowired
-    private SessionRepository sessionRepository;
+    private SessionService sessionService;
 
     @PostMapping("/generate-session")
-    public Map<String, Object> generateSession(@RequestBody GenerateSessionRequest request) {
-        Session session = new Session();
-        session.setId(UUID.randomUUID());
-        session.setUserId(request.getUser_id());
-        session.setCreatedAt(LocalDateTime.now());
-        session.setExpiresAt(LocalDateTime.now().plusHours(1));
-        session.setIsValid(true);
+    public ResponseEntity<Map<String, Object>> generateSession(@RequestBody GenerateSessionRequest request) {
+        String sessionId = sessionService.generateSession(request.getUser_id());
 
-        sessionRepository.save(session);
-
-        Map<String, Object> response = new HashMap<>();
-        response.put("code", 201);
-        response.put("msg", "Sesión creada exitosamente");
-        response.put("session_id", session.getId().toString());
-
-        return response;
+        return ResponseEntity.status(201).body(
+                Map.of(
+                        "code", 201,
+                        "msg", "Sesión creada exitosamente",
+                        "session_id", sessionId
+                )
+        );
     }
 }
