+++ AuthMicroservice Backend/src/main/java/upb/edu/AuthMicroservice/services/SessionService.java
@@ -1,0 +1,16 @@
+package upb.edu.AuthMicroservice.services;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import upb.edu.AuthMicroservice.interactors.GenerateSessionInteractor;
+
+@Service
+public class SessionService {
+
+    @Autowired
+    private GenerateSessionInteractor interactor;
+
+    public String generateSession(int userId) {
+        return interactor.execute(userId).toString();
+    }
+}
